Seperate the code into module
