+++ teacher.rb
@@ -2,6 +2,7 @@ require_relative 'person'
 
 class Teacher < Person
   attr_accessor :specialization
+
   def initialize(age, specialization, name, parent_permission: true)
     super(age, name, parent_permission)
     @specialization = specialization
